Added BIRD-specific metadata and custom licenses
Additional BIRD data:
- coursetype
- courseformat
- selfpaced
- audience
- courseprerequisites

Previously all image license in the MOOChub data were set to the same CC License.
Image licenses are now properly set according to the licenses that were set during file upload. 
An additional Matching-Table was needed to convert the license descriptions from Moodle to descriptions that use spdx names and urls, since this is a requirement of the MOOChub Schema. 
BIRD uses a custom vocabulary for its audience, course types and formats. To manage this vocabulary database tables and settings were added as well.