New settings for vocabularies and providers, generate json on demand ...
## Issue #2: When creating a new set of metadata for a course, form fields that are similar to other fields found in the course settings are automatically prefilled.

## Issue #3: The URLs, that are generated to link images to the Moochub data, now display the images directly, instead of forcing a download. This way Moochub should be able to show the correct overviewimages. 

## Issue #4: In the plugin settings there is now a menu (/local/ildmeta/edit_vocabulary.php) that lets admins edit the vocabularies for the attributes coursetype, courseformat, audience and subjectarea. The different terms are defined by a json array, where every item is an object of the terms in different languages identified by their language code. 
´[
    {
        "de": "Geistes- und Kulturwissenschaften",
        "en": "Humanities and Cultural Studies"
    },
    {
        "de": "Gesundheitswissenschaften",
        "en": "Health Care \/ Health Management"
    },
]´

## Issue #10:
Providers and universities have their own settings menu here "/local/ildmeta/edit_provider.php". Here new providers can be added or existing ones edited or deleted. To create a provider there are now four attributes to set. The name of the provider in German, the name of the provider in English, an URL to a homepage and a image file that represents the logo of the provider. With these additional info added, when generating the Moochub json data, the correct moocProvider, with all its attributes can now be set. 

## Issue #6: The fields in the form to edit metadata are now rearranged so the most important and required fields are on top. The required fields, that can be left blank, are now marked as such and the user will not be able to save the changes, before filling out the required fields.

## Issue #8: The moochub data and bird data are now generated on demand when calling the following URLs: /local/ildmeta/get_moochub_courses.php, /local/ildmeta/get_bird_courses.php. This change was done because the task, that was there before, generated a json file every minute, without checking whether there were any changes made since the last time. Now the data is only generated on demand and definitely contain the most recent metadata.


## Other changes
- Additional field where added to the optional bird section, of the metadata form, to be compatible to the BirdCourses and BirdAcademy requirements. These fields are shortname, abstract, availablefrom and availableuntil
- When generating the bird data, an numerical id (before 'futurelearnlab5', now '5' ) is generated for the individual courses to be compatible to the BirdCourses and BirdAcademy requirements.


This version of the plugin build on the version that is compatible with futurelearnlab. You might experience visual inconsistencies when installing this on other platforms.
When upgrading to this version, I suggest to reinstall the plugin completely since there have been a lot of changes to the database and the upgrade.php script doesn't contain all the routines necessary to make all the necessary changes at this point in time.

The isymetaselect repository will contain a new version aswell, that is compatibel with the changes made in this plugin. Look out for the branch with the same name as this ones.