Optional custom vocabulary, updated default vocab and new metadata
Setting a custom vocabulary is now optional. This can be enabled in the plugins admin settings. If the custom vocabulary is disabled (default), the vocabulary will be updated when upgrading the plugin. This way every site using this plugin and not using  a custom vocabulary, should be generating metadate using the same shared vocabulary, that is dictated by the plugin provider.
Their are now two new select options in the dataform, that can only be set if the bird coursetype is of type "Sprachkurs" or rather the first option in the vocabulary. So the way it is programmewd at the moment this feature will break if a custom vocabulary is used and the first option of the coursetypes is not a type equivalent to "Sprachkurs". The two new datafields describe the type of language course and the language level goal of the course.
Also some minor changes to the language strings (users are addressed with "du" in german and spelliing mistakes in the english lang strings) and additions to the vocabularies ("Business Skills" now called "Professional Skills" and added "Career Skills" to coursetype).