Merged tinjohn/master into origin/master
Deviations to commit a08295 by tinjohn.

Didn't use code to initialize Bird properties. Since this caused override of previously set values when switching the exporttobird value to False. Also this step is redundant since defaults for the bird properties are already set on database level.

The code to update an already existing DB record is now contained in an if-else statement, when a DB record already exists. This makes sure no unnecessary DB calls are made if no update is needed.