Add schema validation and Version Negotiation
JSON result is now validated against the schema, before echoing result. If there is an error in the resulting json the moodle admin is notified by email.

Version negotiation.
By default the newest moochub version is returned (V3). If the client is requesting a specific version that is older than v3, v2 moochub data is generated instead. The content-type header of the response now includes information about the used schema version.