Add event for course metadata update
This commit adds a new event class `ildmeta_updated` to the `local_ildmeta` namespace. This event is triggered when the metadata for a course is updated. The event includes additional information such as the course's indexing status and unique identifier.

The `ildmeta_updated` event is triggered in the `edit_metadata.php` file after the course metadata is successfully saved. The event is created with the course ID, context, and additional information, and then triggered.

This commit also includes language string updates for the German and Dutch translations, adding the translation for the event name "Kursmetadaten aktualisiert" in both languages.

Lastly, the version number of the `local_ildmeta` plugin is updated to `2024081414` in the `version.php` file.